requirements.txt
# Core ML and AI libraries
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# RAG and Vector Database
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.15
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0
beautifulsoup4>=4.12.0

# Text Processing
tiktoken>=0.5.1
nltk>=3.8.1
spacy>=3.7.0

# Web and API
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Science
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.2

# Development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0
jupyter>=1.0.0